+++ jsolex/src/main/resources/whats-new.md
@@ -7,6 +7,7 @@ Here are the new features in this version:
 
 ## Changes in 2.5.4
 
+- Added a spectrum browser
 - Added warning when requested pixel shift isn't available
 - Replace invalid pixel shifts with best fit
 - Display required disk space when generating custom animations